+++ src/RefDocGen/TemplateGenerators/Default/Templates/AssemblyTemplate.razor
@@ -1,4 +1,5 @@
 @using RefDocGen.TemplateGenerators.Shared.TemplateModels.Assemblies
+@using RefDocGen.TemplateGenerators.Shared.TemplateModels.Language
 @using RefDocGen.TemplateGenerators.Shared.TemplateModels.Menu
 @inherits ComponentBase
