+++ src/RefDocGen/TemplateGenerators/Default/Templates/Components/MenuComponent.razor
@@ -1,3 +1,4 @@
+@using RefDocGen.TemplateGenerators.Shared.TemplateModels.Language
 @using RefDocGen.TemplateGenerators.Shared.TemplateModels.Menu
 @using RefDocGen.TemplateGenerators.Shared.TemplateTools
 @using RefDocGen.TemplateGenerators.Shared.Tools
