+++ src/RefDocGen/TemplateGenerators/Default/Templates/Components/NestedTypesSection.razor
@@ -1,3 +1,4 @@
+@using RefDocGen.TemplateGenerators.Shared.TemplateModels.Language
 @inherits ComponentBase
 
 @code {
