+++ src/RefDocGen/TemplateGenerators/Default/Templates/StaticPageTemplate.razor
@@ -1,3 +1,4 @@
+@using RefDocGen.TemplateGenerators.Shared.TemplateModels.Language
 @using RefDocGen.TemplateGenerators.Shared.TemplateModels.Menu
 @inherits ComponentBase
