+++ src/RefDocGen/TemplateGenerators/Shared/TemplateModels/Language/LanguageTM.cs
@@ -1,0 +1,8 @@
+namespace RefDocGen.TemplateGenerators.Shared.TemplateModels.Language;
+
+/// <summary>
+/// Template model representing the .NET language used for displaying syntax.
+/// </summary>
+/// <param name="Name">Name of the language to by displayed.</param>
+/// <param name="Id">Identifier of the language.</param>
+public record LanguageTM(string Name, string Id);
