+++ src/RefDocGen/TemplateGenerators/Shared/TemplateModels/Members/FieldTM.cs
@@ -27,8 +27,8 @@ public record FieldTM(
     string Id,
     string Name,
     TypeLinkTM Type,
-    LocalizedData<string[]> Modifiers,
-    LocalizedData<string>? ConstantValue,
+    LanguageSpecificData<string[]> Modifiers,
+    LanguageSpecificData<string>? ConstantValue,
     AttributeTM[] Attributes,
     string? SummaryDocComment,
     string? RemarksDocComment,
