+++ src/RefDocGen/TemplateGenerators/Shared/TemplateModels/Members/IndexerTM.cs
@@ -44,9 +44,9 @@ public record IndexerTM(
     bool HasGetter,
     bool HasSetter,
     bool IsSetterInitOnly,
-    LocalizedData<string[]> Modifiers,
-    LocalizedData<string[]> GetterModifiers,
-    LocalizedData<string[]> SetterModifiers,
+    LanguageSpecificData<string[]> Modifiers,
+    LanguageSpecificData<string[]> GetterModifiers,
+    LanguageSpecificData<string[]> SetterModifiers,
     AttributeTM[] Attributes,
     string? SummaryDocComment,
     string? RemarksDocComment,
