+++ src/RefDocGen/TemplateGenerators/Shared/TemplateModels/Members/PropertyTM.cs
@@ -50,10 +50,10 @@ public record PropertyTM(
     bool HasGetter,
     bool HasSetter,
     bool IsSetterInitOnly,
-    LocalizedData<string[]> Modifiers,
-    LocalizedData<string[]> GetterModifiers,
-    LocalizedData<string[]> SetterModifiers,
-    LocalizedData<string>? ConstantValue,
+    LanguageSpecificData<string[]> Modifiers,
+    LanguageSpecificData<string[]> GetterModifiers,
+    LanguageSpecificData<string[]> SetterModifiers,
+    LanguageSpecificData<string>? ConstantValue,
     AttributeTM[] Attributes,
     string? SummaryDocComment,
     string? RemarksDocComment,
